feat: обновляет боковое меню и маршруты для управления услугами и персонами с учетом их активности

+++ insideCMS/routes/admin.php
@@ -112,24 +112,27 @@ Route::middleware(['auth', 'verified'])->group(function () use ($modules) {
     });
 
     // Роуты для персон 
+    if ($modules['person'] && $modules['person']['is_active']) {
     Route::controller(PersonaController::class)->group(function () {
         Route::get('persons-admin', 'index')->name('persona-admin');
-        Route::get('add-person', 'create')->name('add-person');
-        Route::post('persons', 'store')->name('persons.store');
-        Route::get('persons/{persona}/edit', 'edit')->name('persons.edit');
-        Route::post('persons/{persona}', 'update')->name('persons.update');
-        Route::delete('persons/{persona}', 'destroy')->name('persons.destroy');
-    });
+            Route::get('add-person', 'create')->name('add-person');
+            Route::post('persons', 'store')->name('persons.store');
+            Route::get('persons/{persona}/edit', 'edit')->name('persons.edit');
+            Route::post('persons/{persona}', 'update')->name('persons.update');
+            Route::delete('persons/{persona}', 'destroy')->name('persons.destroy');
+        });
+    }
 
     // Роуты для услуг
-    Route::controller(ServicesController::class)->group(function () {
+    if ($modules['services'] && $modules['services']['is_active']) {
+        Route::controller(ServicesController::class)->group(function () {
         Route::get('services-admin', 'index')->name('services-admin');
         Route::get('add-services', 'create')->name('add-services');
         Route::post('services', 'store')->name('services.store');
         Route::get('services/{service}/edit', 'edit')->name('services.edit');
         Route::post('services/{service}', 'update')->name('services.update');
         Route::delete('services/{service}', 'destroy')->name('services.destroy');
-
-    });
+        });
+    }
 
  });
